backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aioboto3>=14.1.0",
    "alembic>=1.15.1",
    "asyncpg>=0.30.0",
    "beanie>=1.29.0",
    "colorama>=0.4.6",
    "fastapi>=0.115.12",
    "fastapi-users[beanie,sqlalchemy]>=14.0.1",
    "httpx>=0.28.1",
    "pillow>=11.1.0",
    "pydantic-settings>=2.8.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "sqlalchemy>=2.0.39",
    "uvicorn>=0.34.0",
    "websockets>=15.0.1",
]

[dependency-groups]
dev = [
    "ruff>=0.11.2",
]

[tool.ruff]
lint.select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "I001", # imports
    "I002",
    "FAST001", # fastapi
    "FAST002",
    "FAST003",
    "I", # circular imports
    "C4",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
